% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestry_inference.R
\name{infer_ancestry}
\alias{infer_ancestry}
\title{Infer ancestry}
\usage{
infer_ancestry(vcffile, frqfile = NULL, eps = 1e-04, verbose = F)
}
\arguments{
\item{vcffile}{A \code{character}. Path to an uncompressed vcf file.}

\item{frqfile}{A \code{character}. Path to a frequency file.}

\item{eps}{A \code{numeric}. Bounds minimum and maximum ancestries.}

\item{verbose}{A \code{logical}. If set to TRUE, display additional information.}
}
\description{
Infer ancestry from VCF file
}
